requirements.txt
pyyaml>=6.0
joblib==1.2.0
monty==2022.9.9
numpy==1.23.5
pymatgen==2023.1.30
smol>=0.3.0
sparse-lm>=0.3.0
sympy>=1.11.1
jobflow==0.1.9
atomate2>=0.0.8
scikit-learn==1.2.1
scikit-optimize==0.9.0
scipy==1.10.1
pydantic==1.10.5
polytope==0.2.3
cvxpy==1.3.0
cvxopt
